requirements.txt
aniso8601==7.0.0
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
colorama==0.4.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.5
django-cors-headers==3.7.0
django-filter==2.4.0
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==3.2
iniconfig==1.1.1
oauthlib==3.1.1
packaging==21.0
pluggy==0.13.1
promise==2.3
psycopg2==2.9.1
py==1.10.0
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.2.4
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
Rx==1.6.1
singledispatch==3.6.2
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.6
